Re-organize view helpers functions and classes
